requirements.txt
apturl==0.5.2
async-timeout==4.0.2
attrs==21.2.0
autopep8==2.0.1
bcrypt==3.2.0
beautifulsoup4==4.11.1
blinker==1.4
Brlapi==0.8.3
bs4==0.0.1
cachetools==5.2.0
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
decorator==4.4.2
defer==1.0.6
distro==1.7.0
distro-info===1.1build1
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
duplicity==0.8.21
exceptiongroup==1.1.0
fasteners==0.14.1
future==0.18.2
google==3.0.0
google-api-core==2.11.0
google-api-python-client==2.70.0
google-auth==2.15.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.8.0
googleapis-common-protos==1.57.0
httplib2==0.20.2
huepy==1.2.1
idna==3.3
imageio==2.25.0
imageio-ffmpeg==0.4.8
importlib-metadata==4.6.4
iniconfig==2.0.0
instabot==0.117.0
instagrapi==1.16.31
jeepney==0.7.1
jsonschema==3.2.0
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
mock==5.0.1
monotonic==1.6
more-itertools==8.10.0
moviepy==1.0.3
netifaces==0.11.0
numpy==1.24.2
oauthlib==3.2.0
olefile==0.46
packaging==23.0
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
pluggy==1.0.0
proglog==0.1.10
protobuf==4.21.12
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.20.1
pycodestyle==2.10.0
pycryptodomex==3.15.0
pycups==2.0.1
pydantic==1.10.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.1
PySocks==1.7.1
pytest==7.2.1
python-apt==2.3.0+ubuntu2.1
python-dateutil==2.8.1
python-debian===0.1.43ubuntu1
python-dotenv==0.19.2
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
redis==4.4.0
reportlab==3.6.8
requests==2.25.1
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
responses==0.22.0
rsa==4.9
schedule==1.1.0
SecretStorage==3.3.1
six==1.16.0
soupsieve==2.3.2.post1
spotipy==2.22.0
systemd-python==234
texttable==1.6.4
toml==0.10.2
tomli==2.0.1
tqdm==4.64.1
types-toml==0.10.8.3
typing_extensions==4.4.0
ubuntu-advantage-tools==27.12
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.5
usb-creator==0.3.7
wadllib==1.3.6
websocket-client==1.2.3
xdg==5
xkit==0.0.0
youtube-python==1.0.13
zipp==1.0.0
